requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.0.3
mysql-connector-python
gunicorn
qcloud-python-sts
cos-python-sdk-v5